fix: catch and log event callback errors
Event callbacks are provided by clients and thus can always fail/throw.
Since callback resolution is inside a WS object callback, if an error is
thrown by the callback it'll cause the WS instance to ungracefully shut
down - which causes the socket to silently get stuck.

Catch and log callback execution errors so that doesn't happen.